backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend"
authors = ["Shrey Bhadiyadara <shreybhadiyadara33@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.109.2"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.9"
pinecone-client = "^3.0.2"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
httpx = "^0.28.0"
aiofiles = "^24.1.0"
aiohttp = "^3.9.3"
# google-generativeai = ">=0.8.0,<0.9.0"
langchain-google-vertexai = ">=2,<3"
langchain = ">=0.3.0,<0.4.0"
langchain-community = ">=0.3.0,<0.4.0"
langchain-core = ">=0.3.0,<0.4.0"
langchain-google-genai = ">=2,<3"
tenacity = "^8.2.3"
pinecone-notebooks = "^0.1.0"
langchain-pinecone = ">=0.2,<0.3"
unstructured = "^0.16.11"
pdfminer = "^20191125"
pdfminer-six = "^20240706"
langgraph = "^0.2.70"
pypdf2 = "^3.0.1"
pypdf = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
